% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gems.R
\docType{class}
\name{ArtCohort}
\alias{ArtCohort}
\alias{ArtCohort-class}
\alias{[,ArtCohort-method}
\alias{summary,ArtCohort-method}
\alias{update,ArtCohort-method}
\alias{head,ArtCohort-method}
\alias{tail,ArtCohort-method}
\title{Class \code{"ArtCohort"}}
\arguments{
\item{\code{x}}{object an ArtCohort
\code{i}, \code{j}, \code{drop}  same as for data.frame
\code{...}                       passed on to data.frame method
\code{newsize}                   size of the updated cohort
\code{addbaseline}               baseline for new part of cohort
\code{newInitialStates}          initial states for new part of cohort}
}
\description{
TestIs a S4 class for the artificial cohort generated by \code{simulateCohort}.
}
\section{Slots}{

\describe{
\item{\code{states.number}}{Object of class "numeric": number of states
size             Object of class "numeric": cohort size
baseline         Object of class "matrix": baseline matrix
follow.up        Object of class "numeric": maximum follow-up time
parameters       Object of class "transition.structure": input parameters
parametersCovariances
      Object of class "transition.structure": input covariance matrices
timeToTransition Object of class "matrix": input timeToTransition matrix. logical components
transitionFunctions
      Object of class "transition.structure": input hazard functions
time.to.state    Object of class "data.frame": entry times for each patient into each of the states}
}}

\section{Objects from the Class}{
 Objects are created by calls to the
function \code{simulateCohort}.
}

\examples{

showClass("ArtCohort")

}
\seealso{
\code{\link{simulateCohort}},
\code{\link{transition.structure-class}},
\code{\link{transitionProbabilities}}, {\link{cumulativeIncidence}}
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\keyword{classes}
