% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gems.R
\name{countCost}
\alias{countCost}
\title{Counts a cost of the interventions in the cohort}
\usage{
countCost(cohort, cost)
}
\arguments{
\item{cohort}{a cohort simulated in simulateCohort()}

\item{cost}{a list of lists:stateCost, units and qalys defining costs of basic units of cost, 
cost per state expressed in basic units and QALY for each state.}
}
\value{
\code{c("qaly","cost")}
\code{qaly} an average QALY for an individual in the cohort
\code{cost} an average cost for an individual in the cohort
}
\description{
Counts a cost of the intervention based on the cost data and the times spent in each state.
}
