% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gems.R
\name{generateHazardMatrix}
\alias{generateHazardMatrix}
\title{generate template for transition functions}
\usage{
generateHazardMatrix(statesNumber, statesNamesFrom = paste("State",
  1:statesNumber), statesNamesTo = statesNamesFrom)
}
\arguments{
\item{statesNumber}{the number of states to be considered.}

\item{statesNamesFrom}{names of the states in the model.}

\item{statesNamesTo}{names of the states in the model.}
}
\value{
a \code{transition.structure} of dimension \eqn{N \times N}{N x N},
where \eqn{N} is the number of states and with value "impossible" for all
potential transitions.
}
\description{
This function simplifies generating the matrix of transition functions.
}
\seealso{
\code{\link{transition.structure-class}},
\code{\link{simulateCohort}}
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer, Zofia Baranczuk
}
\keyword{utilities}
