% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gems2.R
\name{makeM}
\alias{makeM}
\title{makeM}
\usage{
makeM(hfNames)
}
\arguments{
\item{hfNames}{- a matrix with names of hazard functions. 
One can use predefined function names: "Impossible", "Exponential", "Weibull", "MultWeibull"}
}
\value{
matrix M generated by generateHazardMatrix
with hazard functions, which names are in hfNames matrix.
}
\description{
Builds hazard function matrix based on the hazard function names matrix. 
If some function is not defined, the name itself is copied to M matrix.
}
\examples{
hfNames <- array(rep("Exponential", 36), dim = c(6,6))
hfNames[3,4:5] <-  rep("impossible",2)
hfNames[1:4,6] <- rep("sr.fun", 4)
hfNames[2, 4:5] <- rep("treat.fun", 2)
hfNames[col(hfNames)<=row(hfNames)]<-"NULL"
M <- makeM(hfNames)
}
