% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gems.R
\docType{methods}
\name{plot,PosteriorProbabilities-method}
\alias{plot,PosteriorProbabilities-method}
\title{plot}
\usage{
\S4method{plot}{PosteriorProbabilities}(x, ci = FALSE, main = paste(x@type,
  "after starting in State", x@states[1], "at time 0"),
  states = 1:dim(x@probabilities)[2], lwd = c(2, 2), col = c("blue",
  "green3"), lty = c(1, 2), xlab = "Time", ylab = "Probability", ...)
}
\arguments{
\item{x}{a cohort simulated by \code{simulateCohort}.}

\item{ci}{a binary variable,  cumulative incidence}

\item{main}{same as in any plot}

\item{states}{names of states in the plot}

\item{lwd}{vector of lentht 2, with first component
for the point estimate and second component
for the confidence interval}

\item{col}{vector of lentht 2, with first component
for the point estimate and second component
for the confidence interval}

\item{lty}{vector of lentht 2, with first component
for the point estimate and second component
for the confidence interval}

\item{xlab}{same as in any plot}

\item{ylab}{same as in any plot}

\item{...}{arguments passed on to main method}
}
\description{
Plots S4 objects
}
