% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gems.R
\docType{class}
\name{PosteriorProbabilities}
\alias{PosteriorProbabilities}
\alias{[,PosteriorProbabilities-method}
\alias{head,PosteriorProbabilities-method}
\alias{tail,PosteriorProbabilities-method}
\title{Class \code{"PosteriorProbabilities"}}
\description{
This S4 class summarizes the posterior probabilities over time for objects
of class \code{"ArtCohort"}
}
\section{Slots}{

\describe{
\item{\code{states}}{Object of class "character": names of states}

\item{\code{times}}{Object of class "numeric": time points at which probabilities are evaluated}

\item{\code{probabilities}}{Object of class "matrix": posterior Probabilities to be in each state at each time}

\item{\code{lower}}{Object of class "matrix": lower prediction bound to be in each state at each time}

\item{\code{upper}}{Object of class "matrix": upper prediction bound to be in each state at each time}

\item{\code{type}}{Object of class "character": describes type of probability}
}}

\section{Objects from the Class}{
 Objects are created by calls to the
function \code{simulateCohort}.
}

\examples{

showClass("PosteriorProbabilities")

}
\seealso{
\code{\link{transitionProbabilities}},
\code{\link{cumulativeIncidence}}, \code{\link{ArtCohort}}
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\keyword{classes}
