% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gems.R
\name{timeInStates}
\alias{timeInStates}
\title{Builds hazard function matrix based on the hazard function names matrix.}
\usage{
timeInStates(mydata, end_time = 100)
}
\arguments{
\item{mydata}{- cohort@time.to.state created in simulateCohort()}

\item{end_time}{- time of the simulation}
}
\value{
time spent in each of the states by each simulated unit
}
\description{
Builds hazard function matrix based on the hazard function names matrix.
}
