% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gems.R
\docType{class}
\name{transition.structure}
\alias{transition.structure}
\alias{transition.structure-class}
\alias{[[,transition.structure-method}
\alias{[[<-,transition.structure-method}
\alias{print,transition.structure-method}
\title{Class \code{"transition.structure"}}
\description{
This S4 class provides a structure to specify different characteristics of
transitions, such as transition functions functions, parameters or parameter
covariances.
}
\section{Objects from the Class}{
 Objects are created by calls to the
functions \code{generateHazardMatrix}, \code{generateParameterMatrix},
\code{generateParameterCovarianceMatrix}.
}

\examples{

  showClass("transition.structure")

}
\seealso{
\code{\link{generateHazardMatrix}},
\code{\link{generateParameterMatrix}},
\code{\link{generateParameterCovarianceMatrix}}
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\keyword{classes}
