% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gems.R
\name{transitionProbabilities}
\alias{transitionProbabilities}
\title{transition probabilities}
\usage{
transitionProbabilities(object, times, M = 100)
}
\arguments{
\item{object}{either the output of \code{\link{simulateCohort}} or the
\code{matrix} with the \code{probabilities} slot of that output.}

\item{times}{a time vector.}

\item{M}{number of groups for calculating confidence intervals.}
}
\value{
an object of class \code{"PosteriorProbabilities"}, containing the
statenames, timepoints and the transition probabilities with pointwise
prediction intervals over time.
}
\description{
Calculates the probabilities and prediction intervals after first state
}
\seealso{
\code{\link{PosteriorProbabilities}},
\code{\link{ArtCohort-class}}, \code{\link{simulateCohort}}
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\keyword{utilities}
